Use mastodon-async from split-entities-crate
This patch changes mastodon-async to be used from my own fork, from the
"split-entities-crate" branch, which splits off the entities module in a
helper crate which can be compiled to WASM.

Because the rest of the mastodon-async crate cannot be compiled to WASM,
but we only need the entities, so this fix is the way to go.

Until the changes are upstream, use my own fork.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>